backend/requirements.txt
flask
flask-cors
sqlalchemy
pdfplumber
pytesseract
Pillow
openai
python-dotenv
requests
